Moved AnimationStation over to our actual code base instead of a library. This allows us to use protobuf, modify the file formats a bit more, and make the code a lot more dynamic.
Also removed .hpp as we do not use that naming convention in our code.
