+++ headers/animationstation/effects/rainbow.h
@@ -1,27 +1,24 @@
-#ifndef _CHASE_H_
-#define _CHASE_H_
+#ifndef _RAINBOW_H_
+#define _RAINBOW_H_
 
-#include "../Animation.hpp"
+#include "Animation.h"
 #include "hardware/clocks.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <vector>
-#include "../AnimationStation.hpp"
+#include "AnimationStation.h"
 
-class Chase : public Animation {
+class Rainbow : public Animation {
 public:
-  Chase(PixelMatrix &matrix);
-  ~Chase() {};
+  Rainbow(PixelMatrix &matrix);
+  ~Rainbow() {};
 
   void Animate(RGB (&frame)[100]);
   void ParameterUp();
   void ParameterDown();
 
 protected:
-  bool IsChasePixel(int i);
-  int WheelFrame(int i);
   int currentFrame = 0;
-  int currentPixel = 0;
   bool reverse = false;
   absolute_time_t nextRunTime = nil_time;
 };
