+++ headers/animationstation/effects/statictheme.h
@@ -7,8 +7,8 @@
 #include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include "../Animation.hpp"
-#include "../AnimationStation.hpp"
+#include "Animation.h"
+#include "AnimationStation.h"
 
 class StaticTheme : public Animation {
 public:
