+++ headers/display/ui/screens/MainMenuScreen.h
@@ -4,7 +4,7 @@
 #include "GPGFX_UI_widgets.h"
 #include "GPGFX_UI_types.h"
 #include "enums.pb.h"
-#include "AnimationStation.hpp"
+#include "AnimationStation.h"
 #include "eventmanager.h"
 
 #define INPUT_MODE_XINPUT_NAME "XInput"
