--- lib/AnimationStation/.editorconfig
@@ -1,17 +1,0 @@
-# EditorConfig is awesome: https://EditorConfig.org
-
-# top-most EditorConfig file
-root = true
-
-[*]
-indent_style = space
-indent_size = 2
-end_of_line = crlf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-
-[*.md]
-indent_style = space
-indent_size = 2
-trim_trailing_whitespace = false
