--- lib/AnimationStation/library.json
@@ -1,13 +1,0 @@
-{
-	"name": "AnimationStation",
-	"version": "0.0.1",
-	"description": "LED Handling for GP2040",
-	"keywords": "c c++ baremetal neopixel led",
-	"authors": [
-		{
-			"name": "Jonathan Barket",
-			"url": "https://jonathanbarket.com"
-		}
-	],
-	"license": "MIT"
-}
