+++ src/animationstation/Effects/CustomThemePressed.cpp
@@ -1,4 +1,4 @@
-#include "CustomThemePressed.hpp"
+#include "CustomThemePressed.h"
 
 std::map<uint32_t, RGB> CustomThemePressed::theme;
