+++ src/animationstation/Effects/StaticTheme.cpp
@@ -1,4 +1,4 @@
-#include "StaticTheme.hpp"
+#include "StaticTheme.h"
 
 std::vector<std::map<uint32_t, RGB>> StaticTheme::themes = {};
